move state logic and fetching data to _app.js and pass through the states
