+++ pages/index.js
@@ -1,31 +1,35 @@
 import styled from 'styled-components';
 import Head from 'next/head';
 import RecordFile from '../components/RecordFile';
-import { useState } from 'react';
-import path from 'path';
-import fsPromises from 'fs/promises';
+// import { useState } from 'react';
+// import path from 'path';
+// import fsPromises from 'fs/promises';
 
-export async function getStaticProps() {
-  const filePath = path.join(process.cwd(), './json/data-dummy.json');
-  const jsonData = await fsPromises.readFile(filePath);
-  const collectionData = JSON.parse(jsonData);
+// export async function getStaticProps() {
+//   const filePath = path.join(process.cwd(), './json/data-dummy.json');
+//   const jsonData = await fsPromises.readFile(filePath);
+//   const collectionData = JSON.parse(jsonData);
 
-  return {
-    props: collectionData,
-  };
-}
+//   return {
+//     props: collectionData,
+//   };
+// }
 
-export default function Home(props) {
-  const collection = props.collection;
+export default function Home({
+  setCollectionState,
+  setSelection,
+  collectionState,
+}) {
+  // const collection = props.collection;
 
-  const myCollection = collection.map((file) => {
-    return { ...file, isChecked: false };
-  });
+  // const myCollection = collection.map((file) => {
+  //   return { ...file, isChecked: false };
+  // });
 
-  const [collectionState, setCollectionState] = useState(myCollection);
-  const [selection, setSelection] = useState([]);
+  // const [collectionState, setCollectionState] = useState(myCollection);
+  // const [selection, setSelection] = useState([]);
 
-  console.log(selection);
+  // console.log(selection);
 
   function handleChange(id) {
     const updatedCollection = collectionState.map((file) => {
